Reorganized types shared between backend and frontend slightly.
